--- projects/telescope-extension/src/app/pages/app.component.html
@@ -1,3 +1,0 @@
-<view-app [extensionNavigations]="config.extension?.navigations">
-  <router-outlet></router-outlet>
-</view-app>
