--- projects/telescope-extension/src/app/pages/app.component.spec.ts
@@ -1,5 +1,0 @@
-describe('test', () => {
-  it('dummy', () => {
-    expect(true).toBeTruthy;
-  });
-});
