+++ projects/telescope-extension/src/app/pages/home/home.module.ts
@@ -1,0 +1,11 @@
+import { HomeModule } from '../../views/home/home.module';
+import { HomeRoutingModule } from './home-routing.module';
+import { HomeComponent } from './home.component';
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+
+@NgModule({
+  declarations: [HomeComponent],
+  imports: [CommonModule, HomeRoutingModule, HomeModule],
+})
+export class AppHomeModule {}
